12. Numbers have life; they're not just symbols on paper

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -8,7 +8,7 @@
 
 void print_number(int n)
 {
-	long nprint;
+	unsigned int nprint;
 
 	if (n < 0)
 	{
